fastapi_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==1.26.4
pillow==10.3.0
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
tensorflow==2.16.1
